requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
python-dateutil==2.8.2
python-dotenv==0.19.1
requests==2.26.0
six==1.16.0
urllib3==1.26.7
pytz==2022.1
munch==2.5.0
PyYAML==6.0